Packages refractoring
