+++ analyzer/restapi-plugin/src/main/java/eu/fasten/analyzer/restapiplugin/mvn/EndpointsListener.java
@@ -1,4 +1,4 @@
-package eu.fasten.analyzer.restapiplugin.api;
+package eu.fasten.analyzer.restapiplugin.mvn;
 
 import org.springframework.context.ApplicationContext;
 import org.springframework.context.ApplicationListener;
