+++ analyzer/restapi-plugin/src/main/java/eu/fasten/analyzer/restapiplugin/mvn/api/impl/EdgeApiServiceImpl.java
@@ -1,20 +1,20 @@
-package eu.fasten.analyzer.restapiplugin.api.mvn.impl;
+package eu.fasten.analyzer.restapiplugin.mvn.api.impl;
 
-import eu.fasten.analyzer.restapiplugin.api.KnowledgeBaseConnector;
-import eu.fasten.analyzer.restapiplugin.api.mvn.FileApiService;
+import eu.fasten.analyzer.restapiplugin.mvn.KnowledgeBaseConnector;
+import eu.fasten.analyzer.restapiplugin.mvn.api.EdgeApiService;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
 @Service
-public class FileApiServiceImpl implements FileApiService {
+public class EdgeApiServiceImpl implements EdgeApiService {
 
     @Override
-    public ResponseEntity<String> getPackageFiles(String package_name,
+    public ResponseEntity<String> getPackageEdges(String package_name,
                                                   String package_version,
                                                   short offset,
                                                   short limit) {
-        String result = KnowledgeBaseConnector.kbDao.getPackageFiles(
+        String result = KnowledgeBaseConnector.kbDao.getPackageEdges(
                 package_name, package_version, offset, limit);
         return new ResponseEntity<>(result, HttpStatus.OK);
     }
